feat: Bump Lucene@8.11.3 and lucene-gosen@8.11.0

+++ languagetool-wikipedia/src/main/java/org/languagetool/dev/index/SearcherResult.java
@@ -35,7 +35,7 @@ public class SearcherResult {
   private int docCount;
   private int maxDocChecked;
   private boolean hasTooManyLuceneMatches;
-  private int luceneMatchCount;
+  private long luceneMatchCount;
   private int skipHits;
   private int numDocs;
 
@@ -81,11 +81,11 @@ public class SearcherResult {
     return hasTooManyLuceneMatches;
   }
 
-  public void setLuceneMatchCount(int luceneMatchCount) {
+  public void setLuceneMatchCount(long luceneMatchCount) {
     this.luceneMatchCount = luceneMatchCount;
   }
 
-  public int getLuceneMatchCount() {
+  public long getLuceneMatchCount() {
     return luceneMatchCount;
   }
